home_work_6 added

+++ posts/views.py
@@ -1,9 +1,13 @@
 from django.shortcuts import render, get_object_or_404, redirect
-from django.http.response import HttpResponse
+from django.http.response import HttpResponse, HttpResponseBadRequest
 from django.http.request import HttpRequest
+from django.contrib.auth.decorators import login_required
+from django.db.models import Q
+
+
 from .models import Post, Tag, Category
 from .forms import PostForm
-from django.db.models import Q
+
 
 def hello_world(request):
     return HttpResponse("<h1>Hello world!</h1>")
@@ -25,8 +29,13 @@ def post_list(request: HttpResponse):
         text = Q(text__icontains=search)
         posts = posts.filter(title | text)
     
-    posts = posts[0:5]
     return render(request, "list_posts.html", {"posts": posts})
+
+
+@login_required
+def my_posts(request: HttpRequest) -> HttpResponse:
+    posts = Post.objects.filter(user=request.user)
+    return render(request, 'post/my_posts.html', {'posts': posts})
 
 
 def post_detail(request, pk):
@@ -36,6 +45,7 @@ def post_detail(request, pk):
     return render(request, 'post/post_detail.html', {'post': post})
 
 
+@login_required
 def create_post(request: HttpRequest) -> HttpResponse:
     if request.method.lower() == 'post':
         form = PostForm(request.POST, request.FILES)
@@ -43,6 +53,7 @@ def create_post(request: HttpRequest) -> HttpResponse:
         if form.is_valid():
             post = form.save(commit=False)
             post.is_published = True
+            post.user = request.user
             post.save()
             form.save_m2m()
 
@@ -64,6 +75,9 @@ def create_post(request: HttpRequest) -> HttpResponse:
 def delete_post(request: HttpRequest, pk: int) -> HttpResponse:
     post = get_object_or_404(Post, id=pk)
     
+    if request.user != post.user:
+        return HttpResponseBadRequest('Not Enough Permission')
+    
     if request.method.lower() == 'post':
         post.delete()
         
@@ -72,3 +86,20 @@ def delete_post(request: HttpRequest, pk: int) -> HttpResponse:
     return render(request, 'post/delete_post.html', context={'post': post})
 
 
+
+def edit_post(request: HttpRequest, pk: int) -> HttpResponse:
+    
+    post = get_object_or_404(Post, id=pk)
+    if post.user != request.user:
+        return HttpResponseBadRequest('Not Enough Permission')
+    
+    form = PostForm(instance=post)
+    
+    if request.method.lower() == 'post':
+        form = PostForm(request.POST, request.FILES, instance=post)
+        
+        if form.is_valid():
+            form.save()
+            return redirect('post_detail', pk=post.pk)
+        
+    return render(request, 'post/edit_post.html', context={'form': form, "post": post, 'categories': Category.objects.all(), 'tags': Tag.objects.all()})
